Fixed failing tests in discretizer, data split and processed set reader
