+++ src/datasets/readers.py
@@ -40,11 +40,13 @@ from pathlib import Path
 import multiprocessing as mp
 from collections.abc import Iterable
 from copy import deepcopy
-from utils import NoopLock, get_iterable_dtype, CustomBins, LogBins
+from metrics import CustomBins, LogBins
 from collections import defaultdict
 from utils.IO import *
 from settings import *
-from utils import zeropad_samples, read_timeseries, is_iterable
+from utils.timeseries import read_timeseries
+from utils.arrays import get_iterable_dtype, is_iterable, zeropad_samples
+from utils.types import NoopLock
 from .mimic_utils import upper_case_column_names, convert_dtype_dict, read_varmap_csv
 from .trackers import ExtractionTracker
 from typing import List, Union, Dict
