+++ src/generators/__init__.py
@@ -8,13 +8,16 @@ from copy import deepcopy
 from pathlib import Path
 from utils.IO import *
 from typing import List, Tuple, Union
-from utils import zeropad_samples, read_timeseries
-from utils import NoopLock, get_iterable_dtype, CustomBins, LogBins
+from utils.timeseries import read_timeseries
+from metrics import CustomBins, LogBins
 from preprocessing.scalers import AbstractScaler
 from datasets.trackers import PreprocessingTracker
 from datasets.readers import ProcessedSetReader
 import multiprocessing as mp
 from pathos.multiprocessing import Pool, cpu_count
+
+from utils.arrays import get_iterable_dtype, zeropad_samples
+from utils.types import NoopLock
 
 
 class AbstractGenerator:
@@ -528,7 +531,7 @@ def process_subject(args: Tuple[List[int], int],
                 y_sample = np.atleast_2d(y_sample).repeat(X_sample.shape[0], axis=0)
                 # .astype(np.int64 if isinstance(y_sample, int) else np.float32)
             else:
-                y_sample = np.atleast_1d(y_sample)
+                y_sample = np.atleast_2d(y_sample)
             X_batch.append(X_sample)
             y_batch.append(y_sample)
             t_batch.append(t_sample)
